# packages/contracts/openapi/bookstore.openapi.json
{
  "components": {
    "schemas": {
...
      "AttachImageDto": {
        "properties": {
          "altText": {
            "maxLength": 250,
            "nullable": true,
            "type": "string"
          },
          "isPrimary": {
            "type": "boolean"
          },
          "sortOrder": {
            "minimum": 0,
            "type": "number"
          },
          "storageKey": {
            "description": "Existing immutable Storage object key produced by upload",
            "format": "uuid",
            "type": "string"
          }
        },
        "required": ["storageKey"],
        "type": "object"
      },
      "AttributeDto": {
        "properties": {
          "key": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "sortOrder": {
            "minimum": 0,
            "type": "number"
          },
          "value": {
            "type": "string"
          }
        },
        "required": ["key", "label", "value"],
        "type": "object"
      },
...
      "CreateProductDto": {
        "properties": {
          "brandId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "categoryId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "description": {
            "description": "Markdown; render with the approved safe renderer",
            "nullable": true,
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "type": "string"
          },
          "referencePrice": {
            "description": "Integer VND amount, at most 14 digits; null means contact for price",
            "nullable": true,
            "type": "string"
          },
          "sku": {
            "maxLength": 64,
            "type": "string"
          },
          "slug": {
            "maxLength": 180,
            "type": "string"
          },
          "summary": {
            "maxLength": 500,
            "nullable": true,
            "type": "string"
          }
        },
        "required": ["sku", "name", "slug"],
        "type": "object"
      },
...
      "ImageMetadataDto": {
        "properties": {
          "altText": {
            "maxLength": 250,
            "nullable": true,
            "type": "string"
          },
          "isPrimary": {
            "type": "boolean"
          },
          "sortOrder": {
            "minimum": 0,
            "type": "number"
          }
        },
        "type": "object"
      },
...
      "MarketplaceButtonResponseDto": {
        "properties": {
          "available": {
            "type": "boolean"
          },
          "marketplace": {
            "enum": ["SHOPEE", "TIKTOK_SHOP"],
            "type": "string"
          },
          "message": {
            "nullable": true,
            "type": "string"
          },
          "url": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": ["marketplace", "url", "available", "message"],
        "type": "object"
      },
      "MarketplaceLinkDto": {
        "properties": {
          "isActive": {
            "type": "boolean"
          },
          "url": {
            "format": "uri",
            "maxLength": 2048,
            "type": "string"
          }
        },
        "required": ["url"],
        "type": "object"
      },
      "MarketplaceLinkResponseDto": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "marketplace": {
            "enum": ["SHOPEE", "TIKTOK_SHOP"],
            "type": "string"
          },
          "productId": {
            "format": "uuid",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        },
        "required": ["id", "productId", "marketplace", "url", "isActive", "createdAt", "updatedAt"],
        "type": "object"
      },
...
      "ProductAttributeResponseDto": {
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "sortOrder": {
            "type": "number"
          },
          "value": {
            "type": "string"
          }
        },
        "required": ["id", "key", "label", "value", "sortOrder"],
        "type": "object"
      },
      "ProductImageResponseDto": {
        "properties": {
          "altText": {
            "nullable": true,
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "isPrimary": {
            "type": "boolean"
          },
          "sortOrder": {
            "type": "number"
          },
          "storageKey": {
            "description": "Admin responses only",
            "type": "string"
          },
          "url": {
            "description": "API-relative URL; apply the configured API mount prefix",
            "type": "string"
          }
        },
        "required": ["id", "altText", "sortOrder", "isPrimary", "url"],
        "type": "object"
      },
      "ProductPageResponseDto": {
        "properties": {
          "hasMore": {
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProductResponseDto"
            },
            "type": "array"
          },
          "limit": {
            "type": "number"
          },
          "page": {
            "type": "number"
          }
        },
        "required": ["items", "page", "limit", "hasMore"],
        "type": "object"
      },
      "ProductResponseDto": {
        "properties": {
          "attributes": {
            "description": "Detail responses only",
            "items": {
              "$ref": "#/components/schemas/ProductAttributeResponseDto"
            },
            "type": "array"
          },
          "brandId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "categoryId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "createdAt": {
            "description": "Admin only",
            "format": "date-time",
            "type": "string"
          },
          "currency": {
            "enum": ["VND"],
            "type": "string"
          },
          "description": {
            "description": "Markdown; detail responses only",
            "nullable": true,
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/ProductImageResponseDto"
            },
            "type": "array"
          },
          "isActive": {
            "description": "Admin only",
            "type": "boolean"
          },
          "marketplaceLinks": {
            "description": "Admin only; includes inactive links",
            "items": {
              "$ref": "#/components/schemas/MarketplaceLinkResponseDto"
            },
            "type": "array"
          },
          "marketplaces": {
            "items": {
              "$ref": "#/components/schemas/MarketplaceButtonResponseDto"
            },
            "type": "array"
          },
          "name": {
            "type": "string"
          },
          "referencePrice": {
            "description": "Integer VND string; null means contact for price",
            "nullable": true,
            "type": "string"
          },
          "sku": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "summary": {
            "nullable": true,
            "type": "string"
          },
          "updatedAt": {
            "description": "Admin only",
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "sku",
          "name",
          "slug",
          "summary",
          "categoryId",
          "brandId",
          "referencePrice",
          "currency",
          "images",
          "marketplaces"
        ],
        "type": "object"
      },
      "PublicationDto": {
        "properties": {
          "isActive": {
            "type": "boolean"
          }
        },
        "required": ["isActive"],
        "type": "object"
      },
      "UpdateAttributeDto": {
        "properties": {
          "key": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "sortOrder": {
            "minimum": 0,
            "type": "number"
          },
          "value": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "UpdateProductDto": {
        "properties": {
          "brandId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "categoryId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "description": {
            "description": "Markdown; render with the approved safe renderer",
            "nullable": true,
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "type": "string"
          },
          "referencePrice": {
            "description": "Integer VND amount, at most 14 digits; null means contact for price",
            "nullable": true,
            "type": "string"
          },
          "sku": {
            "maxLength": 64,
            "type": "string"
          },
          "slug": {
            "maxLength": 180,
            "type": "string"
          },
          "summary": {
            "maxLength": 500,
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "__Secure-bookstore-refresh": {
        "in": "cookie",
        "name": "__Secure-bookstore-refresh",
        "type": "apiKey"
      },
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Product catalogue and singleton-admin content APIs.",
    "title": "Bookstore API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/products": {
      "get": {
        "operationId": "CatalogAdminController_list",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "categoryId",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "brandId",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "isActive",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductPageResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["admin-products"]
      },
      "post": {
        "operationId": "CatalogAdminController_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["admin-products"]
      }
    },
    "/admin/products/{id}": {
      "delete": {
        "operationId": "CatalogAdminController_remove",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["admin-products"]
      },
      "get": {
        "operationId": "CatalogAdminController_detail",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["admin-products"]
      },
      "patch": {
        "operationId": "CatalogAdminController_update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["admin-products"]
      }
    },
    "/admin/products/{id}/attributes": {
      "post": {
        "operationId": "CatalogAdminController_addAttribute",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttributeDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["admin-products"]
      }
    },
    "/admin/products/{id}/attributes/{attributeId}": {
      "delete": {
        "operationId": "CatalogAdminController_removeAttribute",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "attributeId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["admin-products"]
      },
      "patch": {
        "operationId": "CatalogAdminController_updateAttribute",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "attributeId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAttributeDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["admin-products"]
      }
    },
    "/admin/products/{id}/images": {
      "post": {
        "operationId": "CatalogAdminController_attachImage",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttachImageDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["admin-products"]
      }
    },
    "/admin/products/{id}/images/{imageId}": {
      "delete": {
        "operationId": "CatalogAdminController_removeImage",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "imageId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["admin-products"]
      },
      "patch": {
        "operationId": "CatalogAdminController_updateImage",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "imageId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageMetadataDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["admin-products"]
      }
    },
    "/admin/products/{id}/images/{imageId}/content": {
      "get": {
        "operationId": "CatalogAdminController_image",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "imageId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/jpeg": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "image/png": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "image/webp": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["admin-products"]
      }
    },
    "/admin/products/{id}/marketplace-links/{marketplace}": {
      "delete": {
        "operationId": "CatalogAdminController_removeMarketplace",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "marketplace",
            "required": true,
            "schema": {
              "enum": ["SHOPEE", "TIKTOK_SHOP"],
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["admin-products"]
      },
      "put": {
        "operationId": "CatalogAdminController_setMarketplace",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "marketplace",
            "required": true,
            "schema": {
              "enum": ["SHOPEE", "TIKTOK_SHOP"],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarketplaceLinkDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["admin-products"]
      }
    },
    "/admin/products/{id}/publication": {
      "patch": {
        "operationId": "CatalogAdminController_publish",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["admin-products"]
      }
    },
...
    "/products": {
      "get": {
        "operationId": "CatalogPublicController_list",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "categoryId",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "brandId",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductPageResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "tags": ["products"]
      }
    },
    "/products/{slug}": {
      "get": {
        "operationId": "CatalogPublicController_detail",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "tags": ["products"]
      }
    },
    "/products/{slug}/images/{imageId}": {
      "get": {
        "operationId": "CatalogPublicController_image",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "imageId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/jpeg": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "image/png": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "image/webp": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": ""
          }
        },
        "tags": ["products"]
      }
    }
  },
  "servers": [],
...
}